@@ -1,19 +1,21 @@
 {
   "name": "jevyanj",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
+    "babel-plugin-react-require": "^3.1.3",
     "babel-plugin-transform-strict-mode": "^6.24.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "web-vitals": "^2.1.4",
-    "webpack": "^5.68.0",
-    "webpack-dev-server": "^4.7.4"
+    "detect-touch": "^1.1.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.51.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -42,11 +44,12 @@
   "devDependencies": {
     "@babel/plugin-transform-strict-mode": "^7.16.7",
     "babel-core": "^6.26.3",
-    "babel-loader": "^8.2.3",
     "babel-preset-env": "^1.7.0",
-    "concurrently": "^7.0.0",
+    "concurrently": "^7.2.0",
     "less": "^4.1.2",
     "less-watch-compiler": "^1.16.3",
-    "webpack-cli": "^4.9.2"
+    "styled-components": "^5.3.5",
+    "webpack-cli": "^4.9.2",
+    "yarn-upgrade-all": "^0.7.1"
   }
 }
